Avoid verification path.
Since the error is printed in a shared object we have no control over,
the best possible way to deal with this is to avoid verification. The
error gets printed because it is impossible to verify an MLIR module
which has unregistered dialects, because the semantics of unregistered
dialects are unknown. In this case, an SCF::ForOp may be a SymbolTable.
This means that symbols used inside an SCF::ForOp may be redefined by
operations which may be symbol tables.