Provide script to discover dependency versions
The script starts from a given jaxlib release version and attempts to
collect the SHA commit hashes used by this jax release for the MLIR-HLO
and LLVM dependencies.
It uses a mix of searching specific files in the jax/tensorflow repo for
hardcoded hashes and the GitHub API to translate a tensorflow monorepo
hash to an equivalent commit hash in the mlir-hlo mirror repo.

As of now this methods works for jaxlib v0.3.20 - v0.4.6