Always deallocate the register that was allocated
This concerns quantum register allocations in the frontend. Currently we
deallocate the last available alias of an allocated register instead of
the original register reference. Aliases can be created via the
quantum.insert operation.
This however doesn't mesh well with the reasource
allocation/deallocation semantics attached to the quantum
alloc/dealloc ops in MLIR, as the buffer deallocation pass will insert
an additional dealloc op for the original register value, leading to a
double free situation.