[MLIR] Perform buffer deallocation in MLIR on the compiled program (#61)
* Implement the AllocationOpInterface

.. for quantum.alloc. This allows to add the buffer deallocation pass
back into the pass pipeline, as this pass would previously fail on
functions that allocate qubits but do not need them
(required for the adjoint grad method).

* Allow quantum-opt to parse bufferization ops

* Always deallocate the register that was allocated

This concerns quantum register allocations in the frontend. Currently we
deallocate the last available alias of an allocated register instead of
the original register reference. Aliases can be created via the
quantum.insert operation.
This however doesn't mesh well with the reasource
allocation/deallocation semantics attached to the quantum
alloc/dealloc ops in MLIR, as the buffer deallocation pass will insert
an additional dealloc op for the original register value, leading to a
double free situation.

* Check register deallocation in frontend lit tests

* Add additional buffer optimization passes

* Add changelog entry