Separate compatibility matrix workflow
This is required to obtain a separate badge for each configuration
(currently 8: stable/latest for each of Catalyst, PennyLane, and
Lightning).
Each workflow is kept identical with the exception of its configuration
matrix. This facilititates changes to the workflow which can then be
applied identically to each individual workflow file, without needing to
account for small differences between them.