[CI] Produce individual badges for the PL/Lightning compatibility workflow (#87)
* Separate compatibility matrix workflow

This is required to obtain a separate badge for each configuration
(currently 8: stable/latest for each of Catalyst, PennyLane, and
Lightning).
Each workflow is kept identical with the exception of its configuration
matrix. This facilititates changes to the workflow which can then be
applied identically to each individual workflow file, without needing to
account for small differences between them.

* Use reusable workflow & move config files up

GitHub only recognizes workflows right under `/.github/workflows`.

* Temporarily trigger on pull request

* Fix workflow names

* Rename wheel action for better visibility

.. by being higher in the list of workflows.

* Fix naming of stable channel

* Revert "Temporarily trigger on pull request"

This reverts commit 00fe0f4be7dd102a7e81aaf506b66bf2aded3f5b.

* Switch trigger to daily M-F 6:20am EDT

This avoids commits to main showing up as "failing".