JIT compile a wrapper around the generated function.
In order to avoid an exception thrown from the runtime and
unwinding the whole python stack, we need to catch it. A simple
solution in this patch is to create a wrapper around the generated
function that just wraps it around a try-except block.

In order to get around the fact that we do not know the arity of the
generated function at release time, we JIT compile it.

In order to transform C++ exceptions into Python exceptions,
it is necessary to set the global error object in Python via the
Python API. This means that the wrapper is linked against the
Python API, and the CDLL gets transformed into a PyDLL.

Notice the implication that PyDLL will hold the GIL during the
execution of the user compiled function.