Move user container directly under .devcontainers
This is due to a bug in the VS Code not recognizing multiple container
configurations when the container is opened via a URL to the GitHub
repository.

This change would make the user config the default container opened via
links, while the dev config can still be accessed once the repository is
cloned.

See https://github.com/microsoft/vscode-remote-release/issues/8412