Change the ABI to make it statically known AOT. (#106)
* Change the ABI to make it statically known AOT.

When the arity of the function is not known statically,
it is not possible to call a function from C/C++-like languages
without using libffi. To avoid introducing a dependency on libffi,
we fix the arity to be always two.

The first argument is a structure of memrefs which correspond
to the possible return values.

The second argument is a structure of pointers to memrefs which
correspond to the arguments to the functions.