Remove patches.
To remove patches we have copy pasted the python bindings into the
repository. Technically the patches may still be used to avoid having to
retype all necessary changes when the ODS specification changes.
However, they are not necessary in that the changes are maintained and
stored in the repository. So:

1. If one builds from source
2. And does not update the ODS specification with the source

No python bindings are needed.

If the ODS specification changes, then the following steps are required:

1. Generate python bindings
2. Apply patches
3. Test patches against ODS changes (might fail as the patches need to
be updated).

Having to match the ODS changes with every re-generation of the python
bindings is necessary as long as the shared library does not include all
the dialects we want to use.