Add -scalarize and -O0.
Scalarize will turn all tensor<1xT> to tensor<T>. This will avoid
the lowering to linear algebra which will have the impact of reducing
the number of variables seen in the LLVM-IR. It was possible to also use
-linalg-detensorize but -linalg-detensorize actually increased overall
compilation time once we take into account -linalg-detensorize compile
time itself. With -scalarize, we see similar results but without
the compile time increase.

-O0 is added to llc as this will improve compilation times.
The functions that are generated by catalyst (when not using catalyst
control flow operations) will have many variables. From 50-500k. llc
by default uses -O2 which uses a greedy register allocator. Ideally,
we shouldn't need to change from -O2 to -O0 but the large number of
variables is an issue. While we work on reducing the number of
variables in the generated functions (either through supporting
function calls or some other mechanism), we are switching to -O0.
Please note that -O0 was shown to not have an impact on the generated
code as the bottle neck for representative functions is in the
execution of quantum operations. However, this may change as more
and more classical computation happens in the workload.

With -scalarize and -O0 we see a measurable improvement from
~10s to ~4s for compile time in the ChemVQE N=16 (-scalarize
reduces ~13k variables to ~5k variables and -O0 avoids compile time) and
a reduction in execution time from ~191s to ~184s. See PR for more details.