Tighten backprop type declarations & add verifier
For example, the op didn't check whether the declared number of gradient
results matches the number of differentiable arguments, a bug in one of
the bufferization tests.

The `quantum_jacobian` operand name was switched out to `cotangents`
since it more generally describes the purpose of that argument, and in
the way it is used now only receives a slice of the quantum jacobian
anyways.