Remove list case.
This is similar to what is happening in the Hamiltonian, where lists are
no longer allowed. The same is true for JAX where JAX doesn't allow one
to index into a list with a JAX tracer.

x = list[index]

is only allowed when

list is itself a jax.numpy.array or a numpy.array. Not when it is a
list.