Add support for `range` in for loops
A custom range class is added that acts identically to the Python range
class, however it allows tracers to be used as arguments for start,
stop, and step.

The for loop conversion looks for the new range class and can
automatically incorporate it into the iteration bounds of the Catalyst
`for_loop` function.

Some advantages include that a static Python range does not need to be
materialized into a constant array, and dynamic ranges are now also
supported.
A downside is that currently a conversion to Catalyst's for_loop always
takes place when a range is encountered. However, the user may use the
indices obtained from the range to index non-array objects, which is not
detectable by us and would lead to a tracing error.