[Frontend] Add AutoGraph support for Python for loops (#258)
Introduces support for Python `for .. in ...:` statement capture as part
of the compiled program.
Similar to PR #235, AutoGraph is used to convert such statements into
the equivalent Catalyst version before tracing occurs. Specifically, the
following constructs are supported via AutoGraph:
- `for elem in iterable:` - These get converted into a `for_loop(0,
len(iterable), 1)` with `elem = iterable[i]` automatically assigned
using the iteration index, assuming `iterable` is convertible to a JAX
array. If this is not the case, the loop is executed as is in Python.
- `for i in range(start, stop, step):` These are converted directly into
their equivalent `for_loop(start, stop, step)`. Contrary to the default
Python `range`, when AutoGraph is enabled `range` can also accept
dynamic tracers as `start`, `stop`, `step` values. If _any_ exception is
raised during the tracing of the `for_loop` body, Catalyst will fall
back to Python with a warning.
- `for i, elem in enemurate(iterable):` - These get converted into
`for_loop(0, len(iterable), 1)` with the iteration index assigned to the
variable chosen by the user (in this case `i`), and `elem =
iterable[i]`. This also assumes that `iterable` is convertible to an
array, and that the loop body traces without exception, otherwise the
loop is executed in Python.

Note that a warning is raised when when a Python fallback is triggered
due to a tracing exception. Python fallbacks caused by the iterable not
being convertible to array are silent.

[sc-41287]