[Frontend] JAX tracer supporting user-defined tape transformation (#240)
In this PR we aim at making it possible to apply user-defined
transformations to a tape containing Catalyst quantum operators. The
main difficulty is related to the fact that in Catalyst we only trace
programs rather than execute them so the operations on tape: (1) are
filled with abstract parameters, (2) include control flow operations
containing nested tapes.

The approach we are working on would essentially split the tracing
process into classical and quantum parts. In order to do so, we store
JAX tracing contexts and re-enter them as needed. We described the idea
and the associated risks in the [internal
document](https://github.com/XanaduAI/sergei-environment/blob/master/src/python_side_transformations/pst.md).

Tasks:
- [x] Prototyping is done
- [x] Core tracing framework is implemented
- [x] Catalyst extensions are ported to the new tracing framework
  - [x] Control flow ops (For-loops, WhileLoops, Conditionals, Adjoint)
   - [x] CustomOps
   - [x] QubitUnitary
  - [x] Mid-circuit measurement
  - [x] Measurements (Sample, ExpVal, Counts, Probs, State)
- [x] Observables (Compbasis, Paulis, Hadamard,
Hermitian,Tensor,Hamiltonian)
  - [x] Gradient primitives (Grad, JVP, VJP)
- [x] All kinds of execution pipelines are supported
  - [x] Both qjit and qnode
  - [x] Just qjit (no quantum parts)
  - [x] Just qnode (interpretation) 
 - [x] Optimizations are ported
- [x] Qubit insert/extract state preservation (quantum register tracing
optimisation)
- [x] Differentiation-specific procedures (JVP/VJP/Grad) are reviewed
and ported
- [x] All tests pass locally 
- [x] Latest version of Catalyst is merged into the PR
  - [x] Obvious merge conflicts are solves
  - [x] Pytests are fixed
  - [x] Littests are fixed
- [ ] Code is clean
  - [x] Temporary and unused source files are removed
  - [x] JAX internals are collected in a separate utility file
  - [x] Tracing context classes are unified
  - [x] Pylint is happy
  - [x] Docstrings are ported or created
  - [x] CodeFactor is happy
  - [x] Codecov is happy
  - [ ] Remaining FIXMEs are fixed

[sc-41345]

---------

Co-authored-by: Erick <erick.ochoalopez@xanadu.ai>
Co-authored-by: David Ittah <dime10@users.noreply.github.com>