Add enzyme sources to compatibility test matrix. (#285)
**Context:** With the changes in the compiler driver we need access to
Enzyme sources when building the dialects. This is because we need
access Enzyme.h for the forward declaration of `augmentPassBuilder`.

**Description of the Change:** Adds sources to Github Action script.

**Benefits:** Github action passes.

**Possible Drawbacks:** Additional time taken when building tests. We
could instead just do an `extern void
augmentPassBuilder(llvm::PassBuilder &PB); ` and it also works. This
avoids increasing time for Github Actions. I think it might be
worthwhile, but some might consider it not a good practice. Another
alternative is to cache Enzyme sources.