[Frontent] Add quantum control op and the decomposition algorithm (#282)
In this PR we add Catalyst analog of `qml.ctrl` with the similar
interface which supports Catalyst hybrid operations.

There are a number of caveats:

* Currently there is no good way to detect incompatible usage of
abstract wires. In some cases, incorrect usage might result in runtime
failures. Ideally the transformation would emit runtime checks but we
don't have the infrastructure for that at the moment.
* Specifying non-empty work wires leads to a failures on the
lightning-kokkos backend.
* Specifying work wires in conjunction with abstract values for
operation or control wires leads to an infinite recursion in the JAX
tracer.


[sc-40337]

---------

Co-authored-by: David Ittah <dime10@users.noreply.github.com>