[Core] Fix the dependency of catalyst/libs to Z libs (#301)
**Context:** LLVM uses the compression algorithm in various places for
the performance purposes. But it seems not to have many use cases in
MLIR. Building Catalyst MLIR dialects along with the compiler driver
with disabling both `zstd` and `zlib` will reduce the wheel size ~10MB.
Besides, running all integration tests didn't conclude any noticeable
slowdown:
```
With zlib: 73.77user  1:18.83elapsed 
Without zlib: 85.14user 1:18.56elapsed 
```
all being said, disabling `zstd` on M1 machines will not impact the
performance significantly and we may want to consider if asking users to
install `zstd` or `zlib` is out of the scope. I leave this as optional
flags to `mlir/Makefile` so everyone be able to test it.

**Description of the Change:** Building MLIR libraries with disabling
the dependency to `zlib` and `zstd` libraries would remove the
dependency of Catalyst core libs to z libs. Note that it's not
recommended as various tools and features of LLVM rely on the
(de)compression algorithm.

**Benefits:** Fixes #275

[sc-46578]