[MLIR] Extend `adjoint` to support `QubitUnitary` (#304)
**Context:** The adjoint decomposition pass in
`AugmentedCircuitGenerator::generate` makes use of the
`DifferentiableGate` interface to identify parameters that need to be
cached in the forwards pass. The issue is that only gates with
individual float64 parameters are included in this interface, the
QubitUnitary which takes a complex tensor is not.
The adjoint pass should use a new interface ParametrizedGate that
includes any op with numeric parameters (so far the only additional one
would be QubitUnitary).
The difficulty however is caching these more general numeric parameters.
The pass right now makes use of a resizable arraylist for float64
values, and cannot straightforwardly hold other datatypes and tensors
thereof.

**Description of the Change:** As an alternative, instead of storing
arbitary data types in a list, types of `tensor<NxNxcomplex<T>>` are
serialized into `f64` specifically to handle QubitUnitary gates. Since
the `arraylist` is a Last-Input First-Output (LIFO) data structure, one
must pay attention regarding the order of pushing and popping into the
array list and reconstructing the `tensor<NxNxcomplex<T>>` from the
`f64` cache.

**Benefits:** Handles QubitUnitary.

**Possible Drawbacks:** Does not handle arbitrary types.

**Related GitHub Issues:** [sc-46570]

---------

Co-authored-by: David Ittah <dime10@users.noreply.github.com>