[Frontend,MLIR] Improve Compiler driver error reporting (#303)
The goal of this PR is

* [x] Python should report a compiler error, rather than simply
RuntimeError: Compilation failed
* [x] The error message should contain the compilation stage (pass) that
failed.
* [x] The error message should contain MLIR debug info if available,
currently only the MLIR error without context is shown: expected mixed
offsets rank to match mixed sizes rank (0 vs 2) so the rank of the
result type is well-formed.
+ Note: the mentioned error is raised from
[Here](https://github.com/llvm/llvm-project/blob/4706251a3186c34da0ee8fd894f7e6b095da8fdc/mlir/lib/Interfaces/ViewLikeInterface.cpp#L48)
  + Verbose mode now includes MLIR stack traces

In the course of work, the following issues were also found and fixed:

* [x] Fixed CompilerDriver failures on empty pipelines
* [x] Compiler class no longer ignores custom pipelines in textual_ir
mode
* [x] Fixed the name-ordering among the pipeline dump files

This PR adds a small test pass which might be a controversial decision
due to a large number of CMakeLists modifications required. Currently
the pass always fails with an error message - we use it to test the
error reporting. MLIR upstream seems to also practice creating test
passes for dialects, so I think we can re-use it later as well.

[sc-46580]