Remove sanitizer use.
See this error message:

You are trying to dlopen a this-file-does-not-exist.so shared
library with RTLD_DEEPBIND flag which is incompatible with sanitizer
runtime (see https://github.com/google/sanitizers/issues/611 for
details). If you want to run this-file-does-not-exist.so library under
sanitizers please remove RTLD_DEEPBIND from dlopen flags.