Enable building and testing Catalyst with Address Sanitizer (#196)
**Context:** To better catch memory issues and undefined behaviours in
the compiler driver and MLIR Python bindings, all frontend and core
tests are run/checked using the address sanitizer.

**Description of the Change:**
- [X] Enable ASAN for LLVM, MLIR-HLO, Enzyme, Catalyst MLIR Dialects,
and the compile driver
- [X] Fix the the runtime build issue with ASAN
- [X] Add `ENABLE_ASAN=ON/OFF` to the make commands

Note. LLVM is not able to be tested when it's built with GCC and
`ENABLE_ASAN=ON` due to predefined clang specific `LD_PRELOAD` calls
with tests.

Note. We currently build and test the runtime with ASAN and clang in our
GH action. We'll check Catalyst with ASAN in CIs after being able to
build Catalyst from source using clang. (Check #342 for details)

 

**Related GitHub Issues:** Closes #49 

**Ticket:** [sc-41702]

---------

Co-authored-by: David Ittah <dime10@users.noreply.github.com>