Add macros for generating the device entry-point and declarations (#400)
This PR adds a C++ template macro to facilitate the generation of a
device entry-point. For every implementation of `QuantumDevice`, there
are a repeated list of methods declarations in the header file of the
class. These declarations can be templated away to reduce the redundant
code to write a new backend device. This PR addresses this issue with
introducing a couple of macros to generate these declarations. It also
updates the signature of `extern "C"` factory methods to stick to the C
data-types.