Add tests vs pennylane rc branch (#406)
**Context:**
We have this test matrix
[here](https://github.com/PennyLaneAI/plugin-test-matrix#feature-freeze-testing-matrix)
that runs the tests every day for each plugin against the release
candidate branch of Pennylane. It's a way to make sure nothing is broken
between the release branch and the other relevant repositories as we
prepare for the release. Since Catalyst is part of the release now, it
should also be included.

**Description of the Change:**
Added an additional install option for pennylane in the main comparison
workflow for Pennylane, Catalyst and Lightning. It's currently set to
the previous release candidate branch. Also added a workflow with
`pennylane: release-candidate` that runs this.

**Benefits:**
We can add this workflow to the matrix linked above, and easily monitor
compatibility with the release candidate branch and catalyst throughout
feature freeze.

**Possible Drawbacks:**
We will need to update the release candidate branch name in the workflow
on the Monday of feature freeze every release cycle. Also it adds
another CI workflow.