Do not build the runtime with different lightning tags
Lightning tags are now explicitly specified in build recipes to ensure
compatibility with a particular commit in the catalyst repo. Therefor it
makes little sense to test configurations which are known not to work.

The plugin matrix will now only test stable/latest configurations with
respect to the *installed* pennylane-lightning package, not the lightning
sources used to build the runtime.