Build runtime (stable) with Lightning git tag in runtime/Makefile instead of latest_release in the plugin matrix (#422)
Lightning tags have been explicitly specified in build recipes for a
while, to ensure compatibility with a particular commit in the catalyst
repo. Therefor it makes little sense to test configurations which are
known not to work.

The plugin matrix will now build the runtime with a stable lightning
config with respect to the lightning repo tag specified in our build
scripts, which may be either `latest_release` or a commit in the history
of the master branch.