Add CSE pass after lowering from HLO
MLIR generated from the JAX frontend typically contains a lot of
duplicate instructions. For instance, the extraction of a scalar value
from the same scalar tensor could be repeated several times.

Besides making the program larger than it needs to be, it also
makes it more difficult to perform insert/extract elimination
when dynamic wires are used. This was causing the gradient pass to fail
in some instances, specifically when duplicating a hybrid function
while removing all quantum code from it.