Fix leftover quantum operations in classical functions during the gradient passes (#439)
Resolves an issue in the gradient dialect brought to light by #397. The
error raised is as follows:
> error: 'func.func' op cloned during the gradient pass is not free of
quantum ops

The error is thrown when when a quantum function is cloned but with all
of its quantum operations removed, in order to retain only the classical
code. This is done in several places in the gradient dialect.

In this case, certain input programs (see linked issue) would still
cause the gradient pass to be unable to remove all quantum operations
from a function. A snippet of the offending IR would look like:
```mlir
%i0 = tensor.extract %idx : tensor<i64>
%q = quantum.extract %r[%i0]
%i1 = tensor.extract %idx : tensor<i64>
%q = quantum.insert %r[%i1], %q
```
Even though the qubit extraction and insertion are performed at the same
index, the SSA value used in both operations is not the same, and so the
operations are not automatically folded way. This folding is usually
relied upon to remove all quantum operations from a function.

Applying CSE before the gradient pass resolves the issue.

[sc-53218]
fixes #437