Make scipy a pure runtime dependency
This change was prompted by potential mismatches in the openblas library
name at build time vs runtime. This can arise when a different scipy
version is installed on the build vs the user system, because the
library name is specialized with version information.

The specific error fixed is something like:
`ld: cannot find library 'libopenblas<...>.so'`