Provide platform-specific symbol visibility
On Linux, "protected" visibility provides better stack traces while
still preventing symbol conflicts, because the symbol is still exported
yet pre-empting of definitions is disabled. The reason this is necessary
is that the backtracing mechanism on Linux only looks at the dynamic
symbol table.

On macos, protected is not available, but it also doesn't suffer the
platform also doesn't suffer from the same stack trace limitation.