Update MacOS Arm64 workflow to run on Non GitHub Runner (#409)
**Context:** Update Arm64 MacOs workflow to use the new self-hosted Mac
runner

**Description of the Change:**
This PR updates the MacOS workflow to build on the new Self-Hosted Mac
runner on the 24th floor. The GitHub provided XLarge runner is a lot
pricier in comparison to using on-prem infrastructure.

Since self-hosted infrastructure is the exact same machine across job
runs, extra care needs to be taken to ensure that the job does not break
anything on the system level and that files are cleaned up on job
completion. This PR also adds the creation and maintenance of a python
venv where all packages are installed, and then deleted at the end of
the job run.

Another change introduced is the location of the build caches, LLVM
stores full path to files when creating the build, this poses issue when
the value of GITHUB_WORKSPACE changes between jobs. Now a directory is
created within /tmp/ and used for the build. This keeps the path
consistent across GitHub-Hosted/Self-Hosted runner and the within the
Self-Hosted runner if the GITHUB_WORKSPACE changes.

**Benefits:**
- Cost savings as we do not pay for any minutes using our own
infrastructure.
- Do not consume concurrency limit of the PennyLaneAI org when using
self-hosted runners.
- Jobs are not limited to a set time-limit when running.

**Possible Drawbacks:**
- Each job does not run on a clean fresh runner. Therefore, extra care
has to be taken with changes the workflow to ensure system is not
polluted.

If the runner does get into a bad state, we can wipe it and return it to
fresh state if necessary.

**Related GitHub Issues:**
None.

**Evidence of Testing:**
Link to build passing for the temp workflow:
https://github.com/PennyLaneAI/catalyst/actions/runs/7558480273

This build had a cache miss and built all the dependencies.

Link to workflow_dispatch of existing workflow passing:
https://github.com/PennyLaneAI/catalyst/actions/runs/7562797603