Do not specify OpenBLAS library at build time to avoid file name discrepencies (#471)
The OpenBLAS library employed by Catalyst to support JAX `custom_call`
functions is provided by SciPy. Unfortunately, the shared library
distributed by SciPy can have a different file name depending on the
version of SciPy used.

The situation I encountered was having an older version of `scipy`
installed, which provides a library named
`libopenblasp-r0-41284840.3.18.so`. During the frontend installation
with `setup.py`, we compile our `libcustom_calls.cpp` library by linking
against the OpenBLAS library discovered at this time. Additionally,
since `scipy` is listed in the `pyproject.toml` as a build dependency,
pip would actually fetch a different, temporary (and up-to-date) version
of `scipy` for use during the build process. This version shipped with a
shared library named `libopenblasp-r0-23e5df77.3.21.dev.so`. The
difference between these two file names means the loader is unable to
find the BLAS library when opening our custom call library.

To resolve such conflicts, we can avoid providing the OpenBLAS library
at all when compiling `libcustom_calls.cpp`, that is omit
`-lopenblasp-r0-...`, and instead provide it when the user invokes
Catalyst to compile their Python function.

Benefits:
- always use the correct file name for the blas library available on the
user system
- simplify `setup.py` code

Note: Since this is a build error it would be difficult to design a
PyTest for it.

**Tests:**
- linux: https://github.com/PennyLaneAI/catalyst/actions/runs/7642146728
- mac (arm):
https://github.com/PennyLaneAI/catalyst/actions/runs/7642149478
- mac (x86): failing on base branch

[sc-55288]

---------

Co-authored-by: Romain Moyard <rmoyard@gmail.com>