[MLIR][Runtime] Change `__quantum__` to` __catalyst__` for all catalyst defined library. (#464)
**Context:** QIR defines an intermediate representation for quantum
operations. This is done by defining quantum operations as function
calls in LLVM-IR. It is possible to implement quantum operations as
function calls, but this is not the only way to generate a valid
executable. Due to QIR defining quantum operations as functions, it is
very natural to simply provide a definition for these functions and IR
with the library defining these functions. However, if two different
implementers implemented the same functions, then the process would have
issues finding the correct location. There are a couple of ways to solve
this, but one quick solution would just be to use namespaces.

Namespaces in C++ are implemented by namemangling. Since QIR's functions
are C like, there is no name mangling. But we can create a namespace by
just preppending by using a unique name.

**Description of the Change:** Replaces `__quantum__` with
`__catalyst__` in the runtime and generated code for QIR "functions" we
override.

**Benefits:** A program can be linked against catalyst and a second
implementer.

**Possible Drawbacks:** This PR does not rename of all `__quantum__`
functions. We depend on some QIR defined functions. We might want to
avoid relying on other QIR runner libraries to avoid symbol conflicts if
third party libraries redefine these functions. This is not necessarily
a drawback, since it will still be present regardless of whether this PR
is merged or not. But I am pointing that a more complete implementation
of this PR would also get rid of QIR runner library as a dependency.

**Related GitHub Issues:**

[sc-55563]