Sc 55785 enable pr wheel builds (#487)
**Context:** This Pull Request enables a CI feature that will allow
testing of wheel builds as part of Pull Requests.

**Description of the Change:**
This pull request leverages the following GitHub Actions mechanic to
deliver conditionally mandating wheel builds on select PRs:

When a workflow check on a PR is triggered, and all the jobs within the
workflow are skipped, then the workflow itself is marked as skipped.
However, when a workflow is triggered where some jobs are marked as
skipped but others run successfully, then the workflow overall is marked
as success.

There also have been two new labels introduced to facilitate the
requirements.
[requires-wheel-builds](https://github.com/PennyLaneAI/catalyst/labels/requires-wheel-builds)
and
[ci:build-wheels](https://github.com/PennyLaneAI/catalyst/labels/ci%3Abuild-wheels).

`requires-wheel-builds` -> Indicates that the wheel builds MUST pass on
a PR. However, the wheels will not be run on each update (sync) to this
PR. But the wheel builds will rather start, and immediately exit with an
error, thus blocking the PR from being merged.

`ci:build-wheels` -> Indicates that the wheels must be built on each
push (sync) to this PR.

If not labels are missing, then the wheel building workflow is still
triggered, however the wheels themselves are not built (those jobs are
skipped), and the initial job to determine if wheels should be built
returns `false` and then exits as successful, thus the workflow is
marked as success and the PR can be merged even without the wheels being
built against it.


The table below outlines every possible behavior.

| `requires-wheel-builds` | `ci:build-wheels` | Resulting Behavior |
| --- | --- | --- |
| Not Present | Not Present | The wheel jobs are skipped. The overall
wheel workflow exits as successful, the Pull Request can be merged. |
| Present | Not Present | The wheel workflow exits with an error. The
wheels are not built. But the error in the workflow stops the pull
request from being merged with failing tests. |
| Not Present | Present | The wheels are built on each update to the PR.
If the building of the wheels fail, the workflow will fail, but pass
otherwise |
| Present | Present | Same behaviour as previous row |

IMPORTANT NOTE: The `check_if_wheel_build_required` job will need to be
set as a required job to pass on pull requests from the admin panel. The
job itself will use appropriate context to pass fail on PRs.

**Benefits:**
Enable us to make wheel builds conditionally mandatory on PRs.

**Possible Drawbacks:**
External contributors do not have access to assign labels to PRs (even
if they are author). In this case a maintainer would need to add
appropriate label on their behalf.

**Related GitHub Issues:**
None. Shortcut
[sc-55785](https://app.shortcut.com/xanaduai/story/55785/update-catalyst-ci-to-conditionally-build-wheels-on-prs)

---------

Co-authored-by: David Ittah <dime10@users.noreply.github.com>
Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>