[Integration with PL] Change entry point convention (#493)
**Context:** We would like allow a single python package to define
multiple compilers. [Entry points are defined by the following
syntax:](https://packaging.python.org/en/latest/specifications/entry-points/#data-model)

```
name = object reference
```

where:

> The name identifies this entry point within its group. The precise
meaning of this is up to the consumer. For console scripts, the name of
the entry point is the command that will be used to launch it. **Within
a distribution, entry point names should be unique. If different
distributions provide the same name, the consumer decides how to handle
such conflicts.** The name may contain any characters except =, but it
cannot start or end with any whitespace character, or start with [. For
new entry points, it is recommended to use only letters, numbers,
underscores, dots and dashes (regex [\w.-]+).
> 
> The object reference points to a Python object. It is either in the
form importable.module, or importable.module:object.attr. Each of the
parts delimited by dots and the colon is a valid Python identifier.

Previously, in PL the name of an entrypoint was `qjit` and we handled
conflicts between multiple distributions by using the object reference
path to disambiguate between multiple `qjit` definitions. The base of
the object reference path was used as the compiler name.

E.g.,

```
qjit = catalyst:qjit
```

meant that this is the entry point for the `catalyst` compiler.

Names are meant to be unique within a single distribution. It is not
possible to setup a python package with two entry points with the same
name.

E.g.,

```
qjit = foo:qjit
qjit = bar:qjit
```


**Description of the Change:**

Let's change the entrypoint to be:

```
$compiler.qjit = object.path:qjit
```

This allows a package to have multiple `qjit`s defined within a single
distribution.

This commit also takes into account backward compatibility at the moment
of pip configuration. The old naming convention is the default option.
The new naming convention, which will need changes in PL main branch,
can be achieved via:

```
PL_VERSION="==0.35.0.dev0" pip install  -e .
```

**Benefits:** Multiple definitions of different `qjit`'s.

---------

Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>